Renamed files and moved identifier functions to the legacy SpinGroupBase
