+++ src/resonanceReconstruction/rmatrix.hpp
@@ -82,8 +82,8 @@ namespace rmatrix {
   #include "resonanceReconstruction/rmatrix/src/makeCompoundSystem.hpp"
   #include "resonanceReconstruction/rmatrix/src/makeReichMooreChannelData.hpp"
   #include "resonanceReconstruction/rmatrix/src/makeReichMooreCompoundSystem.hpp"
-  #include "resonanceReconstruction/rmatrix/src/makeBreitWignerSpinGroups.hpp"
-  #include "resonanceReconstruction/rmatrix/src/makeBreitWignerCompoundSystem.hpp"
+  #include "resonanceReconstruction/rmatrix/src/makeLegacyBreitWignerSpinGroups.hpp"
+  #include "resonanceReconstruction/rmatrix/src/makeLegacyBreitWignerCompoundSystem.hpp"
   #include "resonanceReconstruction/rmatrix/src/makeLegacyUnresolvedResonanceTable.hpp"
   #include "resonanceReconstruction/rmatrix/src/makeLegacyUnresolvedSpinGroups.hpp"
   #include "resonanceReconstruction/rmatrix/src/makeLegacyUnresolvedCompoundSystem.hpp"
