+++ src/resonanceReconstruction/rmatrix/legacy/resolved/SpinGroup/SingleLevelBreitWigner/src/ctor.hpp
@@ -8,15 +8,6 @@ SpinGroup( Channel< Neutron >&& incident, resolved::ResonanceTable&& table,
            const Energy& qx ) :
   SpinGroupBase( std::move( incident ), std::move( table ) ),
   qx_( qx ),
-  reactions_(
-    [] ( const auto& channel ) -> std::array< ReactionID, 3 > {
-
-      auto incident = channel.particlePair().particle().particleID();
-      auto target = channel.particlePair().residual().particleID();
-      return {{ ReactionID{ incident, target, ReactionType( "elastic" ) },
-                ReactionID{ incident, target, ReactionType( "capture" ) },
-                ReactionID{ incident, target, ReactionType( "fission" ) } }};
-    }( incident ) ),
   elastic_( table.resonances().size() ),
   capture_( table.resonances().size() ),
   fission_( table.resonances().size() ),
