+++ src/resonanceReconstruction/rmatrix/src/makeLegacyBreitWignerSpinGroups.hpp
